requirements.txt
bleak>=0.22.0
requests>=2.31.0
flask>=3.0.0
opencv-python-headless>=4.8.1
paho-mqtt>=1.6.1
python-dotenv>=1.0.0
